refactor index
prepararse para API para bot
